LayoutTests/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/close/close-return-expected.txt

CONSOLE MESSAGE: line 12: WebSocket connection to 'ws://localhost:49001/' failed: WebSocket is closed before the connection is established.

PASS WebSockets: close() return value 
